Add support for local folders (#6)
* Upgrade target framework from 4.5 to 4.7.2

4.5 now shows an "unsupported" warning in VS.

* Implement IVsSolutionEvents7 to handle folder events as well as solutions

Allows using the extensions with "Open local folder" functionality. This works because IVsSolution returns the folder path as solution name. Also check for the existence of the P4CONFIG file in the folder and start in automatic mode if it's present.

* Fix incorrect call to OnAfterOpenSolutionOrFolder instead of OnAfterCloseSolutionOrFolder

* Handle `P4CONFIG` also in the classic VS solution path

Also restructure `SccProvider.ReadUserOptions` to respect the user disabling the P4 connection in preferences.